GUI interface & STM32
STM32
a) MAX31855 was tested and minor fixes were applied
b) The reflow process was completed and it's now ready for former tests
c) The reflow process was also extended in order to handle multiple
thermocouples and SSRs

GUI
a) The communication now handles the most basic commands and it should
now be ready for some former tests
b) A double set of PIDs parameters were set

+++ GUI/GUI_interface_extend.cpp
@@ -4,30 +4,70 @@
 
 // Local defines
 #define MAX_NUM_POINTS      8
+
 #define MAX_TEMPERATURE     300
 #define MIN_TEMPERATURE     25
+#define TEMP_INTERVAL       1
+
+#define TIME_INTERVAL       5
+#define MIN_TIME            0+TIME_INTERVAL
+#define MAX_TIME            1200
+
+#define COEFFS_INTERVAL     0.1
+#define COEFFS_MIN_VALUE    0
+#define COEFFS_MAX_VALUE    10
 
 /*
     Constructor
 */
-GUI_frame_ext::GUI_frame_ext(wxWindow* parent) : GUI_frame( parent )
+GUI_frame_ext::GUI_frame_ext(wxWindow* parent)
+                : GUI_frame( parent )
 {
-    device = new Communication();
+    STM32_device = new Communication();
     // Refresh the choiceBox with a list of available COM ports
     wxCommandEvent dummy;
     this->refresh_COM_list(dummy);
-    // Add a text validator for input filtering
-    this->time_text->SetValidator(wxTextValidator(wxFILTER_NUMERIC));
-    this->temp_text->SetValidator(wxTextValidator(wxFILTER_NUMERIC));
     // Clear the reflow points' list and add the two basic columns
     this->points_list->ClearAll();
     this->points_list->AppendColumn(wxString(wxT("Time [s]")), wxLIST_FORMAT_CENTRE, wxLIST_AUTOSIZE_USEHEADER);
     this->points_list->AppendColumn(wxString(wxT("Temperature [°C]")), wxLIST_FORMAT_CENTRE, wxLIST_AUTOSIZE_USEHEADER);
+
+    // Add the reflow starting point by default
+    this->points_list->InsertItem(0, wxT("0") );
+    this->points_list->SetItem(0, 1, wxT("25"));
+
     // Update the status bar
     this->status_bar->SetStatusText(wxString("Disconnected"));
 
-    // Add the basic layers to the graph
+    // Set allowed temp values into the proper choicebox
+    for (int temp_val=MIN_TEMPERATURE; temp_val<=MAX_TEMPERATURE; temp_val+=TEMP_INTERVAL) {
+        this->temp_choice->Append( wxString::Format(wxT("%d"), temp_val) );
+    }
+    this->temp_choice->SetSelection(0);
 
+    // Set allowed time values into the proper choicebox
+    for (int time_val=MIN_TIME; time_val<=MAX_TIME; time_val+=TIME_INTERVAL) {
+        this->time_choice->Append( wxString::Format(wxT("%d"), time_val) );
+    }
+    this->time_choice->SetSelection(0);
+
+    // Set predefined P-I-D values
+    for (float coeffs=COEFFS_MIN_VALUE; coeffs<=COEFFS_MAX_VALUE; coeffs+=COEFFS_INTERVAL) {
+        this->P_choice_1->Append( wxString::Format(wxT("%.3f"), coeffs) );
+        this->I_choice_1->Append( wxString::Format(wxT("%.3f"), coeffs) );
+        this->D_choice_1->Append( wxString::Format(wxT("%.3f"), coeffs) );
+        this->P_choice_2->Append( wxString::Format(wxT("%.3f"), coeffs) );
+        this->I_choice_2->Append( wxString::Format(wxT("%.3f"), coeffs) );
+        this->D_choice_2->Append( wxString::Format(wxT("%.3f"), coeffs) );
+    }
+    this->P_choice_1->SetSelection(0);
+    this->I_choice_1->SetSelection(0);
+    this->D_choice_1->SetSelection(0);
+    this->P_choice_2->SetSelection(0);
+    this->I_choice_2->SetSelection(0);
+    this->D_choice_2->SetSelection(0);
+
+    // Add the basic layers to the graph
     m_plot = new mpWindow( this, -1, wxDefaultPosition, wxDefaultSize, wxDOUBLE_BORDER );
 	m_plot->AddLayer( new mpScaleX( wxT("Elapsed time [s]")));
     m_plot->AddLayer( new mpScaleY( wxT("Temperature [°C]")));
@@ -70,7 +110,7 @@ void GUI_frame_ext::connect_to_COM( wxCommandEvent& event )
 {
     wxString portname = this->m_choice1->GetString(this->m_choice1->GetSelection());
 
-    if( device->connect(portname) != 0 ) {
+    if( STM32_device->connect(portname) != 0 ) {
         wxMessageBox( wxT("Device not found"), wxT("Error"), wxICON_INFORMATION);
         this->status_bar->SetStatusText(wxString("Disconnected"));
     }else{
@@ -92,17 +132,11 @@ void GUI_frame_ext::insert_point( wxCommandEvent& event )
         return;
     }
 
-    // Get data from input boxes
-    if ( get_time_temp_data(time, temperature) < 0 ){
-        return;
-    }
-
-    // Check that the specified temperature is inside allowed ranges
-    if ( (temperature>MAX_TEMPERATURE)||(temperature<MIN_TEMPERATURE) ) {
-        wxMessageBox( wxString::Format("Temperature out of range!\nValid range is between %d and %d", MIN_TEMPERATURE, MAX_TEMPERATURE),
-                        wxT("Error"), wxICON_INFORMATION);
-        return;
-    }
+    // Get the data from the text boxes
+    wxString time_text = this->time_choice->GetString(this->time_choice->GetSelection());
+    time_text.ToLong(&time);
+    wxString temp_text = this->temp_choice->GetString(this->temp_choice->GetSelection());
+    temp_text.ToLong(&temperature);
 
     // check for the right position in which the new point should be inserted
     int index = 0;
@@ -125,14 +159,9 @@ void GUI_frame_ext::insert_point( wxCommandEvent& event )
         }
     }
 
-    // Convert numeric valuese into string again
-    std::ostringstream time_text, temp_text;
-    time_text << time;
-    temp_text << temperature;
-
     // If we get here then the user specified valid data, so we can add them to our list
-    points_list->InsertItem(index, time_text.str() );
-    points_list->SetItem(index, 1, temp_text.str());
+    this->points_list->InsertItem(index, time_text );
+    this->points_list->SetItem(index, 1, temp_text );
 
     // Refresh the graph
     this->update_graph();
@@ -144,37 +173,11 @@ void GUI_frame_ext::insert_point( wxCommandEvent& event )
 void GUI_frame_ext::clear_list( wxCommandEvent& event )
 {
     points_list->DeleteAllItems();
-}
-
-/*
-    Returns the time and temperature data contained into the text boxes (if valid)
-*/
-int GUI_frame_ext::get_time_temp_data(long& time, long& temperature)
-{
-    // Get the data from the text boxes
-    wxString time_text_data = this->time_text->GetLineText(0);
-    wxString temp_text_data = this->temp_text->GetLineText(0);
-
-    // First check that the text boxes are not empty ...
-    if ( time_text_data.length() == 0 ){
-        wxMessageBox( wxT("Time text box is empty"), wxT("Error"), wxICON_INFORMATION);
-        return -1;
-    }
-    if ( temp_text_data.length() == 0 ){
-        wxMessageBox( wxT("Temperature text box is empty"), wxT("Error"), wxICON_INFORMATION);
-        return -2;
-    }
-
-    // ... and contain valid data
-    if( !time_text_data.ToLong(&time) ){
-        wxMessageBox( wxT("Invalid time value"), wxT("Error"), wxICON_INFORMATION);
-        return -3;
-    }
-    if( !temp_text_data.ToLong(&temperature) ){
-        wxMessageBox( wxT("Invalid temperature value"), wxT("Error"), wxICON_INFORMATION);
-        return -4;
-    }
-    return 0;
+    // Keep the reflow starting point
+    this->points_list->InsertItem(0, wxT("0") );
+    this->points_list->SetItem(0, 1, wxT("25"));
+    // Refresh the graph
+    this->update_graph();
 }
 
 /*
@@ -182,12 +185,6 @@ int GUI_frame_ext::get_time_temp_data(long& time, long& temperature)
 */
 void GUI_frame_ext::remove_point( wxCommandEvent& event )
 {
-    // Check if the list is already empty or not
-    if ( this->points_list->GetItemCount() == 0 ) {
-        wxMessageBox( wxT("The list is already empty"), wxT("Error"), wxICON_INFORMATION);
-        return;
-    }
-
     // Check that a line is selected from the points list
     if ( this->points_list->GetSelectedItemCount() == 0 ) {
         wxMessageBox( wxT("No line selected from the list"), wxT("Error"), wxICON_INFORMATION);
@@ -196,7 +193,70 @@ void GUI_frame_ext::remove_point( wxCommandEvent& event )
 
     // Get the index of the selected line
     int selected_item = this->points_list->GetNextItem(-1, wxLIST_NEXT_ALL, wxLIST_STATE_SELECTED);
+    // Of course the default starting point added by default cannot be cancelled
+    if ( selected_item == 0 ) {
+        wxMessageBox( wxT("The first default point cannot be cancelled"), wxT("Error"), wxICON_INFORMATION);
+        return;
+    }
+
+    // Remove the point from the list
     this->points_list->DeleteItem(selected_item);
+
+    // Refresh the graph
+    this->update_graph();
+}
+
+/**
+ *  1) Send data points to the STM32
+ *  2) Sends PIDs parameters to the STM32
+ *  3) Starts the relow process
+ */
+void GUI_frame_ext::start( wxCommandEvent& event )
+{
+    // Clear the reflow list which might be loaded into the STM32
+    STM32_device->clear_reflow_list();
+
+    // Send data points to the STM32
+    int num_items = this->points_list->GetItemCount();
+    int index;
+    wxListItem current_item;
+    wxString time, temperature;
+
+    for (index=0; index<num_items; index++){
+        current_item.SetId(index);
+
+        current_item.SetColumn(0);
+        this->points_list->GetItem(current_item);
+        time = current_item.GetText();
+        current_item.SetColumn(1);
+        this->points_list->GetItem(current_item);
+        temperature = current_item.GetText();
+
+        STM32_device->add_reflow_point(time, temperature);
+    }
+
+    // Send PIDs parameters
+    wxString P_coeff, I_coeff, D_coeff;
+    P_coeff = this->P_choice_1->GetString(this->P_choice_1->GetSelection());
+    I_coeff = this->I_choice_1->GetString(this->I_choice_1->GetSelection());
+    D_coeff = this->D_choice_1->GetString(this->D_choice_1->GetSelection());
+    STM32_device->set_PID_parameters(1, P_coeff, I_coeff, D_coeff);
+
+    P_coeff = this->P_choice_2->GetString(this->P_choice_2->GetSelection());
+    I_coeff = this->I_choice_2->GetString(this->I_choice_2->GetSelection());
+    D_coeff = this->D_choice_2->GetString(this->D_choice_2->GetSelection());
+    STM32_device->set_PID_parameters(2, P_coeff, I_coeff, D_coeff);
+
+    // Start the reflow process
+    STM32_device->start_reflow_process();
+}
+
+/**
+ *  Stops the relow process
+ */
+void GUI_frame_ext::stop( wxCommandEvent& event )
+{
+    STM32_device->stop_reflow_process();
 }
 
 /**
